pyproject.toml
[tool.poetry]
name = "shoes-store"
version = "0.1.0"
description = ""
authors = ["Veronelli <facu2000veronelli@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.27.1"
fastapi = "^0.109.2"
pytest-mock = "^3.12.0"
strawberry-graphql = "^0.221.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pre-commit = "^3.6.2"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
